Add files via upload
This is the overarching gesture system. Unfortunately this is not a complete runnable project because of a python3.12 vs. 3.13 versioning issue on my laptops architecture. As a result media pipe cannot be coexisting with the cflib. To combat this issue I used interprocess communication. So you'll see the ipc files in both this and in the c/p for reference [5] that were used to circumvent this issue by running two different python distributions.